refactor(metrics): getMetricsRawData を移動
